modify inventory

+++ doc/inventory_sp/3_sp_stockout.sql
@@ -7,7 +7,7 @@ IN  inventory_sn     int(11)   ,
 IN  batch     varchar(100)  ,
 IN  warehouse varchar(100)  ,
 IN  position  varchar(100) ,
-IN  quantity   int(10) ,
+IN  quantity   double(16,3) ,
 IN	irradiation     varchar(100)   ,
 IN  tpc     varchar(100)    ,
 IN  vendor_id     int(11)   ,
@@ -19,38 +19,14 @@ IN  remark 		  varchar(500)
 
 BEGIN
 
-INSERT INTO inv_inventory_out
-(	item_code , inventory_sn     , batch     ,warehouse , position  ,  quantity    ,irradiation      , tpc      ,
-  vendor_id   ,customer_id   ,status       ,stockout_date ,
-  create_time, create_by,
-  update_by,
-	update_time,
-  remark 		 )
-VALUES (
-  item_code ,
-  inventory_sn ,
-  batch     ,
-  warehouse ,
-  position  ,
-  quantity    ,
-	irradiation      ,
-  tpc      ,
-  vendor_id   ,
-  customer_id   ,
-	0      ,
-	stockout_date ,
-  sysdate() ,
-  create_by,
-  create_by   ,
-	sysdate() ,
-  remark
-);
+INSERT INTO inv_inventory_out( item_code , inventory_sn , batch , warehouse , position  , quantity  , irradiation , tpc ,
+                                vendor_id , customer_id , status ,stockout_date , create_time, create_by , update_by , update_time , remark )
+              VALUES ( item_code , inventory_sn , batch , warehouse , position , quantity ,	irradiation , tpc ,
+                        vendor_id , customer_id ,	0 ,	stockout_date , sysdate() , create_by , create_by , sysdate() , remark );
+
 END
 ;;
 DELIMITER ;
-
-
-
 
 
 DROP PROCEDURE IF EXISTS sp_stockOut;
@@ -61,7 +37,7 @@ IN  item_code  varchar(100) ,
 IN  batch     varchar(100)  ,
 IN  warehouse varchar(100)  ,
 IN  position  varchar(100) ,
-IN  var_quantity   int(10) ,
+IN  var_quantity   double(16,3) ,
 IN	irradiation     varchar(100)   ,
 IN  tpc     varchar(100)    ,
 IN  vendor_id     int(11)   ,
@@ -86,7 +62,7 @@ start transaction;
 update inv_inventory  set quantity=res_quantity   where sn =inventory_sn;
 
 call sp_insert_inv_out(item_code, inventory_sn, batch, warehouse, position ,
-var_quantity ,irradiation ,tpc  , vendor_id , customer_id , stockout_date , create_by , remark 	 	);
+                      var_quantity ,irradiation ,tpc , vendor_id , customer_id , stockout_date , create_by , remark 	 	);
 
 
 commit;
@@ -94,8 +70,6 @@ commit;
 END
 ;;
 DELIMITER ;
-
-
 
 
 
@@ -108,9 +82,9 @@ IN  var_sn  int(11)
 BEGIN
 
 declare tmp_inventory_sn int(11);
-declare tmp_quantity int(10);
-declare res_quantity INT default 0;
-declare inventory_quantity INT default 0;
+declare tmp_quantity double(16,3);
+declare res_quantity double(16,3) default 0;
+declare inventory_quantity double(16,3) default 0;
 
 select inventory_sn, quantity into tmp_inventory_sn, tmp_quantity from inv_inventory_out where sn=var_sn;
 select  quantity into inventory_quantity from inv_inventory where sn=tmp_inventory_sn;
